fmt: Move FormatEvent into format module
